+++ src/api/endpoints/collections/get-collection/v2.ts
@@ -16,7 +16,7 @@ export const getCollectionV2Options: RouteOptions = {
   tags: ["api", "x-deprecated"],
   plugins: {
     "hapi-swagger": {
-      order: 3,
+      deprecated: true,
     },
   },
   validate: {
@@ -37,7 +37,7 @@ export const getCollectionV2Options: RouteOptions = {
     schema: Joi.object({
       collection: Joi.object({
         id: Joi.string(),
-        slug: Joi.string().description("Open Sea slug"),
+        slug: Joi.string().allow(null, "").description("Open Sea slug"),
         name: Joi.string().allow(null, ""),
         metadata: Joi.object().allow(null),
         sampleImages: Joi.array().items(Joi.string().allow(null, "")),
@@ -117,9 +117,9 @@ export const getCollectionV2Options: RouteOptions = {
         ownerCount: Joi.number(),
         attributes: Joi.array().items(
           Joi.object({
-            key: Joi.string().allow(null),
-            kind: Joi.string().allow(null),
-            count: Joi.number().allow(null),
+            key: Joi.string().allow(null, ""),
+            kind: Joi.string().allow(null, ""),
+            count: Joi.number().allow(null, ""),
           })
         ),
       }).allow(null),