+++ src/api/endpoints/execute/get-execute-bid/v2.ts
@@ -37,10 +37,10 @@ const version = "v2";
 export const getExecuteBidV2Options: RouteOptions = {
   description: "Create bid (offer)",
   notes: "Generate a bid and submit it to multiple marketplaces",
-  tags: ["api", "Orderbook"],
+  tags: ["api", "x-deprecated"],
   plugins: {
     "hapi-swagger": {
-      order: 11,
+      deprecated: true,
     },
   },
   validate: {