+++ src/api/endpoints/orders/get-orders-bids/v1.ts
@@ -366,7 +366,7 @@ export const getOrdersBidsV1Options: RouteOptions = {
           metadata: r.metadata,
           source: {
             id: source?.address,
-            name: source?.name,
+            name: source?.metadata.title || source?.name,
             icon: source?.metadata.icon,
             url: source?.metadata.url,
           },