+++ src/api/endpoints/orders/index.ts
@@ -2,6 +2,7 @@ export * from "@/api/endpoints/orders/get-order-executed/v1";
 export * from "@/api/endpoints/orders/get-orders/v1";
 export * from "@/api/endpoints/orders/get-orders/v2";
 export * from "@/api/endpoints/orders/get-orders-all/v1";
+export * from "@/api/endpoints/orders/get-orders-all/v2";
 export * from "@/api/endpoints/orders/get-orders-bids/v1";
 export * from "@/api/endpoints/orders/get-orders-bids/v2";
 export * from "@/api/endpoints/orders/get-orders-bids/v3";