+++ src/api/endpoints/tokens/get-tokens-details/v2.ts
@@ -25,7 +25,6 @@ export const getTokensDetailsV2Options: RouteOptions = {
   tags: ["api", "x-deprecated"],
   plugins: {
     "hapi-swagger": {
-      order: 22,
       deprecated: true,
     },
   },
@@ -402,7 +401,7 @@ export const getTokensDetailsV2Options: RouteOptions = {
               validUntil: r.floor_sell_value ? r.floor_sell_valid_until : null,
               source: {
                 id: source?.address,
-                name: source?.name,
+                name: source?.metadata.title || source?.name,
                 icon: source?.metadata.icon,
                 url: source?.metadata.url,
               },