+++ src/api/endpoints/tokens/get-tokens-details/v3.ts
@@ -436,7 +436,7 @@ export const getTokensDetailsV3Options: RouteOptions = {
               validUntil: r.floor_sell_value ? r.floor_sell_valid_until : null,
               source: {
                 id: source?.address,
-                name: source?.name,
+                name: source?.metadata.title || source?.name,
                 icon: source?.metadata.icon,
                 url: source?.metadata.url,
               },