+++ src/models/daily-volumes/daily-volume.ts
@@ -595,7 +595,7 @@ export class DailyVolume {
       queries.push({
         query: `
             UPDATE collections
-            SET day${period}_floor_sell_value = $/floor_sell_value/,
+            SET day${period}_floor_sell_value = $/floor_sell_value/${valuesPostfix},
                 updated_at = now()                              
             WHERE id = $/collection_id/`,
         values: row,
@@ -604,13 +604,10 @@ export class DailyVolume {
 
     try {
       await idb.none(pgp.helpers.concat(queries));
-    } catch (e: any) {
+    } catch (error) {
       logger.error(
         "daily-volumes",
-        JSON.stringify({
-          msg: `Error while updating the floor_sell_value of period ${period} in the collections table`,
-          exception: e.message,
-        })
+        `Error while updating the floor_sell_value${valuesPostfix} of period ${period} in the collections table: ${error}`
       );
       return false;
     }