+++ src/sync/events/data/index.ts
@@ -16,6 +16,7 @@ import * as wyvernV2 from "@/events-sync/data/wyvern-v2";
 import * as wyvernV23 from "@/events-sync/data/wyvern-v2.3";
 import * as x2y2 from "@/events-sync/data/x2y2";
 import * as zeroExV4 from "@/events-sync/data/zeroex-v4";
+import * as cryptoPunks from "@/events-sync/data/cryptopunks";
 
 // All events we're syncing should have an associated `EventData`
 // entry which dictates the way the event will be parsed and then
@@ -60,7 +61,9 @@ export type EventDataKind =
   | "element-erc1155-sell-order-filled"
   | "element-erc1155-buy-order-filled"
   | "quixotic-order-filled"
-  | "nouns-auction-settled";
+  | "nouns-auction-settled"
+  | "cryptopunks-punk-bought"
+  | "cryptopunks-transfer";
 
 export type EventData = {
   kind: EventDataKind;
@@ -111,6 +114,8 @@ export const getEventData = (eventDataKinds: EventDataKind[] | undefined) => {
       element.erc1155SellOrderFilled,
       quixotic.orderFulfilled,
       nouns.auctionSettled,
+      cryptoPunks.punkBought,
+      cryptoPunks.transfer,
     ];
   } else {
     return (
@@ -201,6 +206,10 @@ const internalGetEventData = (kind: EventDataKind): EventData | undefined => {
       return quixotic.orderFulfilled;
     case "nouns-auction-settled":
       return nouns.auctionSettled;
+    case "cryptopunks-punk-bought":
+      return cryptoPunks.punkBought;
+    case "cryptopunks-transfer":
+      return cryptoPunks.transfer;
     default:
       return undefined;
   }